refactor: move soil metadata out of soil id directory
