+++ dev-client/src/model/soilMetadata/soilMetadataFunctions.ts
