+++ dev-client/src/screens/LocationScreens/components/LocationSoilIdCard.tsx
@@ -35,10 +35,10 @@ import {
 } from 'terraso-mobile-client/components/NativeBaseAdapters';
 import {SoilIdStatusDisplay} from 'terraso-mobile-client/components/SoilIdStatusDisplay';
 import {SoilIdStatus} from 'terraso-mobile-client/model/soilId/soilIdSlice';
-import {findSelectedMatch} from 'terraso-mobile-client/model/soilId/soilMetadataFunctions';
-import {useSoilIdSelection} from 'terraso-mobile-client/model/soilId/soilMetadataHooks';
 import {useSoilIdData} from 'terraso-mobile-client/model/soilIdMatch/soilIdMatchHooks';
 import {getTopMatch} from 'terraso-mobile-client/model/soilIdMatch/soilIdRanking';
+import {findSelectedMatch} from 'terraso-mobile-client/model/soilMetadata/soilMetadataFunctions';
+import {useSoilIdSelection} from 'terraso-mobile-client/model/soilMetadata/soilMetadataHooks';
 
 type LocationSoilIdCardProps = {
   coords: Coords;
