+++ dev-client/src/store/reducers.ts
@@ -30,8 +30,8 @@ import {reducer as preferencesReducer} from 'terraso-mobile-client/model/prefere
 import projectReducer from 'terraso-mobile-client/model/project/projectSlice';
 import siteReducer from 'terraso-mobile-client/model/site/siteSlice';
 import soilIdReducer from 'terraso-mobile-client/model/soilId/soilIdSlice';
-import soilMetadataReducer from 'terraso-mobile-client/model/soilId/soilMetadataSlice';
 import soilIdMatchReducer from 'terraso-mobile-client/model/soilIdMatch/soilIdMatchSlice';
+import soilMetadataReducer from 'terraso-mobile-client/model/soilMetadata/soilMetadataSlice';
 import syncReducer from 'terraso-mobile-client/model/sync/syncSlice';
 
 const sliceReducers = {
